update: libraries

+++ data/storage/smbj/src/main/java/com/wa2c/android/cifsdocumentsprovider/data/storage/smbj/SmbjProxyFileCallbackSafe.kt
@@ -42,7 +42,7 @@ class SmbjProxyFileCallbackSafe(
     override fun onWrite(offset: Long, size: Int, data: ByteArray): Int {
         if (mode != AccessMode.W) { throw ErrnoException("Writing is not permitted", OsConstants.EBADF) }
         return processFileIo(coroutineContext) {
-            file.write(data, offset, 0, size)
+            file.write(data, offset, 0, size).toInt()
         }
     }
